+++ database/migrations/2016_08_29_173100_remove_individual_players_etc.php
@@ -53,35 +53,46 @@ class RemoveIndividualPlayersEtc extends Migration
                 //all players (not individual) are set to go, now we can remove the column that is not need it anymore.
 
             }
-            Schema::connection('mysql_champ')->table('players', function ($table) {
-                $table->dropForeign('players_team_id_foreign');
-                $table->dropColumn('team_id');
+            if (Schema::connection('mysql_champ')->hasColumn('players', 'team_id')) {
+                Schema::connection('mysql_champ')->table('players', function ($table) {
+                    $table->dropForeign('players_team_id_foreign');
+                    $table->dropColumn('team_id');
+                });
+            }
+        }
+
+        if (Schema::connection('mysql_champ')->hasTable('individual_players')) {
+//            dd("here is the problem");
+            $indPlayer = \App\Models\Championship\IndividualPlayer::all()->toArray();
+
+            $this->back_individualPlayers_table($indPlayer);
+            foreach ($indPlayer as &$p) {
+//                dd($p);
+                $string = 'none_yet';// this is to create a string that would allow team mates to get assign to a team if they didnt have one
+                $game_id = $p['game_id'];
+                $tournaments = \App\Models\Championship\Tournament::where('game_id', $p['game_id'])->get()->toArray();
+                unset($p['id']);
+                unset($p['game_id']);
+                $newPlayer = \App\Models\Championship\Player::create($p);
+                //create the pivot table info for players and tournaments
+                foreach ($tournaments as $tournament) { // this is a bold fix because we just go with every torunament that is for that game, even if the user didnt sign for it, but because there is only one game hopefully will work and we wont need it again :)
+                    \App\Models\Championship\Players_Tournaments::create(['player_id' => $newPlayer->id, 'tournament_id' => $tournament['id']]);
+                }
+            }
+            //all individual players are in the player table and in the pivot table for tournaments now we can drop the table
+
+        }
+        if (Schema::connection('mysql_champ')->hasColumn('individual_players', 'game_id')) {
+            Schema::connection('mysql_champ')->table('individual_players', function (Blueprint $table) {
+                $table->dropForeign(['game_id']);
+                $table->dropColumn(['game_id']);
             });
         }
 
         if (Schema::connection('mysql_champ')->hasTable('individual_players')) {
-            $indPlayer = \App\Models\Championship\IndividualPlayer::all()->toArray();
-            if (isset($indPlayer[0]['game_id'])) {
-
-                $this->back_individualPlayers_table($indPlayer);
-                foreach ($indPlayer as &$p) {
-//                dd($p);
-                    $string = 'none_yet';// this is to create a string that would allow team mates to get assign to a team if they didnt have one
-                    $game_id = $p['game_id'];
-                    $tournaments = \App\Models\Championship\Tournament::where('game_id', $p['game_id'])->get()->toArray();
-                    unset($p['id']);
-                    unset($p['game_id']);
-                    $newPlayer = \App\Models\Championship\Player::create($p);
-                    //create the pivot table info for players and tournaments
-                    foreach ($tournaments as $tournament) { // this is a bold fix because we just go with every torunament that is for that game, even if the user didnt sign for it, but because there is only one game hopefully will work and we wont need it again :)
-                        \App\Models\Championship\Players_Tournaments::create(['player_id' => $newPlayer->id, 'tournament_id' => $tournament['id']]);
-                    }
-                }
-                //all individual players are in the player table and in the pivot table for tournaments now we can drop the table
-
-            }
-            Schema::connection('mysql_champ')->drop('individual_players');
+            Schema::connection('mysql_champ')->dropIfExists('individual_players');
         }
+
 
     }
 
@@ -92,17 +103,24 @@ class RemoveIndividualPlayersEtc extends Migration
      */
     public function down()
     {
-        Schema::connection('mysql_champ')->create('individual_players', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('username')->unique();
-            $table->string('email')->unique();
-            $table->string('name');
-            $table->string('phone');
-            $table->timestamps();
-        });
-        Schema::connection('mysql_champ')->table('players', function ($table) {
-            $table->integer('team_id')->default(0);
-        });
+
+        if (!Schema::connection('mysql_champ')->hasTable('individual_players')) {
+            Schema::connection('mysql_champ')->create('individual_players', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('username')->unique();
+                $table->string('email')->unique();
+                $table->string('name');
+                $table->string('phone');
+                $table->integer('game_id')->references('id')->on('games');;
+                $table->timestamps();
+            });
+        }
+        if (!Schema::connection('mysql_champ')->hasColumn('players', 'team_id')) {
+            Schema::connection('mysql_champ')->table('players', function ($table) {
+                $table->unsignedInteger('team_id')->default(0);
+                $table->foreign('team_id')->references('id')->on('teams');
+            });
+        }
     }
 
     /**
